demo/requirements.txt
transformers
torch
asyncpraw
fastapi[all]
python-dotenv
